Move defs into seperate file
