+++ plugins/connection/v1.py
@@ -100,7 +100,7 @@ options:
 """
 
 
-from ansible_collections.ansible.snmp.plugins.plugin_utils.netsnmp_wrapper import (
+from ansible_collections.ansible.snmp.plugins.plugin_utils.netsnmp_defs import (
     Snmpv1Connection,
 )
 from ansible_collections.ansible.snmp.plugins.plugin_utils.snmp_connection_base import (
