+++ plugins/connection/v2c.py
@@ -86,7 +86,7 @@ options:
 """
 
 
-from ansible_collections.ansible.snmp.plugins.plugin_utils.netsnmp_wrapper import (
+from ansible_collections.ansible.snmp.plugins.plugin_utils.netsnmp_defs import (
     Snmpv2cConnection,
 )
 from ansible_collections.ansible.snmp.plugins.plugin_utils.snmp_connection_base import (
